+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Calendar.js - Change Log:
 
 ### Version 0.9.1:
+- Added new public function updateEventDateTimes().
+- Drag & Drop now uses the new updateEventDateTimes() function (much faster).
 - 
 
 ### Version 0.9.0: